split db modal file
